requirementb.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
email-validator
fpdf2
